requirements.txt
anyio==4.8.0
certifi==2025.1.31
charset-normalizer==3.4.1
construct==2.10.68
construct-typing==0.5.6
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jsonalias==0.1.1
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
sniffio==1.3.1
solana==0.36.6
solders==0.26.0
typing_extensions==4.12.2
urllib3==2.3.0
websockets==15.0
